WORKAROUND : assignment submission data loss when using htmlspecialcharacters
This is a workaround for a bug where htmlspecialchars (< > &) typed in the
AJAX form input result in ALL text afterward being deleted by moodle,
prior to saving to database. AFAIKT this bug has has something todo with
moodle's AJAX text input sanitizers acting differently than the the
default moodle form submission sanitizers. This hack works around the
problem simply by un-encoding the characters on client side, (except for
"&", which needs to be replaced completely with "and"). This causes the
moodle sanitizers to skip NON HTML use cases of these characters.

This workaround DOES NOT address the continued problem of when moodle sanitizes
the submission. Any characters removed in this case, results in ALL text
after the offending bits to be removed also. Such a case can be
reproduced by submitting an assignment with any "hand typed" HTML.